added snap turns and teleportation
snap turning at 45 degrees with a delay of 0.05 seconds. currently the entire ground is a teleportation area since there is no need to block off any spaces and no stationary objects to which a person should be able to teleport. when the chest model is added we will add a teleportation anchor in front of it that faces the chest.